+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - dotnet: emit API features for objects created via the newobj instruction #1186 @mike-hunhoff
 - dotnet: emit API features for generic methods #1231 @mike-hunhoff
 - Python 3.11 support #1192 @williballenthin
+- dotnet: emit calls to/from MethodDef methods #1236 @mike-hunhoff
 
 ### Breaking Changes